% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logloss.R
\docType{data}
\name{log.loss}
\alias{log.loss}
\title{Log Loss mlr custom measure}
\format{An object of class \code{Measure} of length 10.}
\usage{
log.loss
}
\description{
Log Loss mlr custom measure
}
\examples{

## Define tasks
iris$Species <- as.integer(iris$Species == "virginica")
lrn = makeLearner("classif.nnet", predict.type = "prob")
iris.task <- makeClassifTask(data = iris, target = "Species")
## Train a learner and make predictions
mod = train(learner = lrn, iris.task)
pred = predict(mod, task = iris.task)
## Calculate the performance using the new measure
performance(pred, measures = log.loss)

}
\keyword{datasets}

