% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logloss.R
\name{logLoss}
\alias{logLoss}
\title{Function to calculate the Log Loss Measure}
\usage{
logLoss(task, model, pred, feats, extra.args)
}
\arguments{
\item{task}{task}

\item{model}{model}

\item{pred}{pred}

\item{feats}{feats}

\item{extra.args}{extra.args}
}
\description{
Function to calculate the Log Loss Measure
}

